Move platform-utils to @geti/core
