+++ web_ui/packages/core/src/platform-utils/dto/utils.interface.ts
