+++ web_ui/packages/core/src/platform-utils/hooks/use-gpu-provider.hook.ts
