+++ web_ui/packages/core/src/platform-utils/hooks/use-platform-utils.hook.ts
@@ -1,14 +1,14 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
-import { useFeatureFlags } from '@geti/core/src/feature-flags/hooks/use-feature-flags.hook';
-import { useApplicationServices } from '@geti/core/src/services/application-services-provider.component';
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 import { useAuth } from 'react-oidc-context';
 import { v4 as uuid } from 'uuid';
 
-import QUERY_KEYS from '../../../../packages/core/src/requests/query-keys';
+import { useFeatureFlags } from '../../feature-flags/hooks/use-feature-flags.hook';
+import QUERY_KEYS from '../../requests/query-keys';
+import { useApplicationServices } from '../../services/application-services-provider.component';
 import { ProductInfoEntity, WorkflowId } from '../services/utils.interface';
 
 const placeholderUuid = uuid();
