+++ web_ui/packages/core/src/platform-utils/services/create-api-platform-utils-service.ts
@@ -1,10 +1,9 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
-import { apiClient } from '@geti/core';
-
-import { CreateApiService } from '../../../../packages/core/src/services/create-api-service.interface';
-import { API_URLS } from '../../../../packages/core/src/services/urls';
+import { apiClient } from '../../client/axios-instance';
+import { CreateApiService } from '../../services/create-api-service.interface';
+import { API_URLS } from '../../services/urls';
 import { Environment, ProductInfoEntityDTO } from '../dto/utils.interface';
 import { PlatformUtilsService, ProductInfoEntity } from './utils.interface';
