+++ web_ui/packages/core/src/platform-utils/services/create-in-memory-platform-utils-service.ts
