+++ web_ui/packages/core/src/platform-utils/services/utils.interface.ts
