+++ web_ui/packages/core/src/services/get-api-services.ts
@@ -19,7 +19,6 @@ import { createApiMediaService } from '../../../../src/core/media/services/api-m
 import { createApiModelsService } from '../../../../src/core/models/services/api-models-service';
 import { createApiOrganizationsService } from '../../../../src/core/organizations/services/api-organizations-service';
 import { createApiPersonalAccessTokensService } from '../../../../src/core/personal-access-tokens/personal-access-tokens-service';
-import { createApiPlatformUtilsService } from '../../../../src/core/platform-utils/services/create-api-platform-utils-service';
 import { createApiProjectService } from '../../../../src/core/projects/services/api-project-service';
 import { createApiDatasetStatisticsService } from '../../../../src/core/statistics/services/api-dataset-statistics-service';
 import { createApiModelStatisticsService } from '../../../../src/core/statistics/services/api-model-statistics-service';
@@ -28,6 +27,7 @@ import { createApiSupportedAlgorithmsService } from '../../../../src/core/suppor
 import { createApiTestsService } from '../../../../src/core/tests/services/api-tests-service';
 import { createApiUserSettingsService } from '../../../../src/core/user-settings/services/api-user-settings-service';
 import { createApiFeatureFlagService } from '../feature-flags/services/api-feature-flag-service';
+import { createApiPlatformUtilsService } from '../platform-utils/services/create-api-platform-utils-service';
 import { createApiOnboardingService } from '../users/services/api-onboarding-service';
 import { createApiUsersService } from '../users/services/api-users-service';
 import { createApiWorkspacesService } from '../workspaces/services/api-workspaces-service';
